requirements.txt
pydantic
protobuf
starlette
python-dotenv
firebase-admin
gspread
fastapi
uvicorn[standard]
google-cloud-firestore